changed the location of _fast_kde function from kdeplot.py to plot_utils.py
