+++ arviz/plots/backends/matplotlib/violinplot.py
@@ -5,8 +5,7 @@ import numpy as np
 from . import backend_show
 from ....stats import hpd
 from ....stats.stats_utils import histogram
-from ...kdeplot import _fast_kde
-from ...plot_utils import get_bins, make_label, _create_axes_grid
+from ...plot_utils import get_bins, make_label, _create_axes_grid, _fast_kde
 
 
 def plot_violin(
