+++ arviz/plots/loopitplot.py
@@ -5,8 +5,11 @@ from matplotlib.colors import to_rgb, rgb_to_hsv, hsv_to_rgb, to_hex
 from xarray import DataArray
 
 from ..stats import loo_pit as _loo_pit
-from .plot_utils import _scale_fig_size, get_plotting_function
-from .kdeplot import _fast_kde
+from .plot_utils import (
+    _scale_fig_size,
+    get_plotting_function,
+    _fast_kde,
+)
 
 
 def plot_loo_pit(
