+++ arviz/plots/plot_utils.py
@@ -1,6 +1,7 @@
 """Utilities for plotting."""
 import warnings
 from itertools import product, tee
+from scipy.signal import gaussian, convolve
 import importlib
 
 import packaging
@@ -8,11 +9,11 @@ import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib as mpl
 import xarray as xr
+from ..stats.stats_utils import histogram
 from scipy.stats import mode
 
 
 from ..utils import conditional_jit
-from .kdeplot import _fast_kde
 from ..rcparams import rcParams
 
 
@@ -697,6 +698,7 @@ def calculate_point_estimate(point_estimate, values, bw):
     point_value : float
         best estimate of data distribution
     """
+    point_value = None;
     if point_estimate == "auto":
         point_estimate = rcParams["plot.point_estimate"]
     elif point_estimate not in ("mean", "median", "mode", None):
@@ -718,3 +720,71 @@ def calculate_point_estimate(point_estimate, values, bw):
         point_value = np.median(values)
 
     return point_value
+
+
+def _fast_kde(x, cumulative=False, bw=4.5, xmin=None, xmax=None):
+    """Fast Fourier transform-based Gaussian kernel density estimate (KDE).
+
+    The code was adapted from https://github.com/mfouesneau/faststats
+
+    Parameters
+    ----------
+    x : Numpy array or list
+    cumulative : bool
+        If true, estimate the cdf instead of the pdf
+    bw : float
+        Bandwidth scaling factor for the KDE. Should be larger than 0. The higher this number the
+        smoother the KDE will be. Defaults to 4.5 which is essentially the same as the Scott's rule
+        of thumb (the default rule used by SciPy).
+    xmin : float
+        Manually set lower limit.
+    xmax : float
+        Manually set upper limit.
+
+    Returns
+    -------
+    density: A gridded 1D KDE of the input points (x)
+    xmin: minimum value of x
+    xmax: maximum value of x
+    """
+    x = np.asarray(x, dtype=float)
+    x = x[np.isfinite(x)]
+    if x.size == 0:
+        warnings.warn("kde plot failed, you may want to check your data")
+        return np.array([np.nan]), np.nan, np.nan
+
+    len_x = len(x)
+    n_points = 200 if (xmin or xmax) is None else 500
+
+    if xmin is None:
+        xmin = np.min(x)
+    if xmax is None:
+        xmax = np.max(x)
+
+    assert np.min(x) >= xmin
+    assert np.max(x) <= xmax
+
+    log_len_x = np.log(len_x) * bw
+
+    n_bins = min(int(len_x ** (1 / 3) * log_len_x * 2), n_points)
+    if n_bins < 2:
+        warnings.warn("kde plot failed, you may want to check your data")
+        return np.array([np.nan]), np.nan, np.nan
+
+    _, grid, _ = histogram(x, n_bins, range_hist=(xmin, xmax))
+
+    scotts_factor = len_x ** (-0.2)
+    kern_nx = int(scotts_factor * 2 * np.pi * log_len_x)
+    kernel = gaussian(kern_nx, scotts_factor * log_len_x)
+
+    npad = min(n_bins, 2 * kern_nx)
+    grid = np.concatenate([grid[npad:0:-1], grid, grid[n_bins : n_bins - npad : -1]])
+    density = convolve(grid, kernel, mode="same", method="direct")[npad : npad + n_bins]
+    norm_factor = (2 * np.pi * log_len_x ** 2 * scotts_factor ** 2) ** 0.5
+
+    density /= norm_factor
+
+    if cumulative:
+        density = density.cumsum() / density.sum()
+
+    return density, xmin, xmax
